AY-9: InputStream changed to Reader
