+++ ReaderHw/src/Main.java
@@ -6,7 +6,7 @@ import java.io.*;
 public class Main {
     public static void main(String[] args) throws IOException {
         File file = new File("src/file.txt");
-        InputStream is = new LowerToUpperAndUppertoLowerInputStream(new FileInputStream(file));
+        Reader is = new LowerToUpperAndUppertToLowerReader(new FileReader(file));
         int c;
         while ((c = is.read()) != -1){
             System.out.print((char)c);
