+++ ReaderHw/src/file.txt
@@ -1,1 +1,2 @@
 UPPERCASE lowercase
+Азат Алексеевич
